Final_grad Doctor

+++ app/src/main/java/com/example/finalproj_doctor/Model/Doctor.java
@@ -1,7 +1,5 @@
 package com.example.finalproj_doctor.Model;
 
-import com.example.finalproj_doctor.Model.Location;
-
 import java.util.ArrayList;
 
 public class Doctor {
@@ -9,7 +7,7 @@ public class Doctor {
     private Location location;
     private ArrayList<String> speciality;
     private String photo , homeConsultation , accountStatus , _id , address , description , email , name , phone , slug;
-    private int averageRating;
+    private float averageRating;
 
     public Location getLocation() {
         return location;
@@ -107,11 +105,11 @@ public class Doctor {
         this.slug = slug;
     }
 
-    public int getAverageRating() {
+    public float getAverageRating() {
         return averageRating;
     }
 
-    public void setAverageRating(int averageRating) {
+    public void setAverageRating(float averageRating) {
         this.averageRating = averageRating;
     }
